Add utility to merge po files
Surprisingly I could not find a tool that merges updated translations into a
.po file in a way that

a) merges only missing translations (does not overwrite existing ones)
b) produces a minimal diff so that there is a chance to manually check the result

Therefore I wrote my own. You can use it for a single language

python development/tools/mergepo.py ../lyx-2.1/po fr

or for all languages:

python development/tools/mergepo.py ../lyx-2.1/po

The python code is not the most elegant one, but it works. What is missing is
a command line switch to merge changed translations as well. This is useful
for languages that have not yet received any translation update in the
development branch, only in the stable branch: In this case we know that
translations that are not identical in both branches should be overtaken from
the stabkle one.