Cmake export tests: Allow deeper label nesting.
Previously all labels got the depth '7' while processing 'suspiciousTestss'.
The depth is used to sort how our labels are used to build a test-label.

Say a test gets label a, b, and c, with depth 3, 8 and 2.
The constructed test-label will be "b:a:c"
For new (to be implemented) 'reason-labels' it is more convenient to assign them higher values.