Explain why we do not update the local layout
We do not update the local layout of our .lyx files because users
would then not be able to export to older formats. For example, if a
2.2.0 user exported a template to 2.1.x format and tried to open the
file in LyX 2.1.x, there would be an error because the file would
contain a local layout whose format is too new. The root reason for
this is that we do not convert layouts to older layout formats.