Do not setup building of 3rdparty/libiconv
This causes build problems with automake < 1.14 because two files have the same name. This only works well when using subdir-objects, which we only do after 1.14.

We do not want to build libiconv on target which support autotools anyway, but rely on a built-in one.

This fixes building on oldish ubuntu 12.04.