Fix line endings
The default behaviour of git for text files is to store unix line endings in
the repository, and to convert to native line endings in the local workspace.
This plays well with the LyX behaviour of storing files in native line endings.
Unfortunately it may break if files are sent around by email and submitted on
a different OS than the one they were last saved, and this did probably happen
in 6b0632eea.