Remove stmary10.sfd from source package
We do now have .sfd sources for all fonts, so either we distribute all of them
with the source package or none. Since the accumulated size is not negligible,
and the are not needed to build LyX, but serve a really special purpose, we
do not add them to the source package for now.