Document need to use same subset if --rerun-failed
When running the tests with ctest, --rerun-failed can be used to run
the tests that failed on the previous run. However, the same subset
must be specified because it is the test numbers that are used to
index which tests failed on the previous run. For example, it would
be incorrect to do the following:

  ctest -R export
  ctest --rerun-failed

The following is correct:

  ctest -R export
  ctest -R export --rerun-failed