LyXToolBox: a QToolBox with minimum size management
The purpose of this custom widget is to allow the use of a QToolBox in a limited
area. The stock QToolBox does not provide a minimum size hint that depends on
the size of the pages; it assumes that there is enough room.  This subclass sets
the minimal size of the QToolbox. Without this, the size of the QToolbox is only
determined by values in the ui file and therefore causes portability and
localisation issues. Note that the computation of the minimum size hint depends
on the minimum size hints of the page widgets. Therefore page widgets must have
a layout with layoutSizeContraint = SetMinimumSize or similar.

(cherry picked from commits 576cb5df77604b50175b2410ac006c5c5c110e2b
                            7e9d0a938bb02510304822f6bce4df3bcfe4bfc9
                            efbc297b955da58cef7bf2d63dfbc8b2942c671e)