Latest brazilian translations from Georger
After some discussion between Georger, Pedro, Pavel and me Georger came to the
conclusion that Mapa is not the correct translation of Chart in this context.
Originally he proposed Gráfico, but since that would be identical to the
translation of Graph we decided that Diagram is better. It is a bit more
general than Gráfico, but it is also used in pt_PT, and the german translation
uses the german equivalent, which is more general as well. Those who need a
Gráfico can still use Graph.

Unfortunately Geoger based the new .po file on the remerged one at
259196e1a6c54, so we have now the situation that pt_BR is remerged, but the
other languages are not.