Remerge strings except for sv and fr
This is a complete remerge of strings (done with cmake) except for fr.po and
sv.po. Those files had already been remerged after the last string changes,
but they use a different sorting, so I did not want to produce an unneeded
diff.
The differences in the comment lines are a well known problem we have: The
last remerge was done on windows, mine was done on linux.