More fixes for #9897 (wrong inset-icon size in PDF)
Inset-icons are auto-scaled according to their width.
This is problematic, if converted from SVG
with the bounding box set to content (ignoring the page size)
and the content is not square.

OTOH, ignoring page size is done for good reasons - we usually want just
the drawing area when inserting an image into a document.

An invisible square with fixes this problem.

Some icons had spurious invisible elements outside the page -- leading
to too small results in the PDF. These were removed.